thesis-experiment/backend/requirements.txt
Django==2.0.1
djangorestframework==3.7.7
django-autoslug-iplweb==1.9.4.dev0
django-oauth-toolkit==1.0.0
django-cors-headers==2.1.0